+++ src/components/info-dialog-props.ts
@@ -1,7 +1,7 @@
 import { t } from "@lingui/macro";
 
 import { InfoDialogButtonProps } from "src/components/info-dialog";
-import { WikiPageSlug } from "src/domain/wiki";
+import { WikiPageSlug } from "src/domain/types";
 
 export const infoDialogProps = () =>
   ({